pyproject.toml
[tool.poetry]
name = "Clux"
version = "0.1.0"
description = ""
authors = ["Pablo Martin PErez"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.1"
peewee = "^3.13.1"
python-dotenv = "^0.10.5"
pytest = "^5.3.5"
coverage = "^5.0.3"
flask_babelex = "^0.9.4"
flask_images = "^3.0.2"
wtforms = "^2.2.1"
flask-admin = "^1.5.5"
flask-mail = "^0.9.1"
wtf-peewee = "^3.0.0"
flask-restful = "^0.3.8"
flask-security-too = "^3.3.3"
bcrypt = "^3.1.7"
flask-wtf = "^0.14.3"
flask-httpauth = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
